fix(cursor): extract AppStatePaths for nonisolated cache directory
CompanionStore.defaultURL() is MainActor-isolated; share PTB_STATE_DIR resolution via AppStatePaths for Cursor API disk cache.
