+++ Sources/PokeTokenBar/Core/AppStatePaths.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+/// Application Support state directory for PokeTokenBar files.
+/// `PTB_STATE_DIR` overrides the default for development/QA isolation.
+enum AppStatePaths {
+    static func directory() -> URL {
+        let override = (ProcessInfo.processInfo.environment["PTB_STATE_DIR"] ?? "")
+            .trimmingCharacters(in: .whitespacesAndNewlines)
+        let dir: URL
+        if !override.isEmpty {
+            dir = URL(fileURLWithPath: override, isDirectory: true)
+        } else {
+            dir = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask)[0]
+                .appendingPathComponent("PokeTokenBar")
+        }
+        try? FileManager.default.createDirectory(at: dir, withIntermediateDirectories: true)
+        return dir
+    }
+}
